Avoid bad truncation of 64-bit values during masking

+++ src/platform.h
@@ -238,12 +238,12 @@ HEDLEY_WARNING("NEON disabled due to missing arm_neon.h header");
 #elif defined(_ISOC11_SOURCE)
 	// C11 method
 	// len needs to be a multiple of alignment, although it sometimes works if it isn't...
-	#define ALIGN_ALLOC(buf, len, align) *(void**)&(buf) = aligned_alloc(align, ((len) + (align)-1) & ~((align)-1))
+	#define ALIGN_ALLOC(buf, len, align) *(void**)&(buf) = aligned_alloc(align, ((len) + (align)-1) & ~(size_t)((align)-1))
 	#define ALIGN_FREE free
 #elif defined(__cplusplus) && __cplusplus >= 201700
 	// C++17 method
 	#include <cstdlib>
-	#define ALIGN_ALLOC(buf, len, align) *(void**)&(buf) = std::aligned_alloc(align, ((len) + (align)-1) & ~((align)-1))
+	#define ALIGN_ALLOC(buf, len, align) *(void**)&(buf) = std::aligned_alloc(align, ((len) + (align)-1) & ~(size_t)((align)-1))
 	#define ALIGN_FREE free
 #else
 	#define ALIGN_ALLOC(buf, len, align) if(posix_memalign((void**)&(buf), align < sizeof(void*) ? sizeof(void*) : align, (len))) (buf) = NULL
